feat(front) : notchBox finalization
