+++ packages/scss/src/components/index.scss
@@ -51,4 +51,4 @@
 @forward 'grid'; // 40 Ko
 @forward 'tableSticked'; // 67 Ko
 @forward 'timepicker'; //
-@forward 'notchedBox';
+@forward 'notchBox';
