+++ packages/scss/src/components/notchBox/exports.scss
