--- packages/scss/src/components/notchedBox/index.scss
@@ -1,22 +1,0 @@
-@use 'exports' as *;
-
-.notchedBox {
-	@include vars;
-	@include component;
-
-	&.mod-top {
-		@include top;
-	}
-
-	&.mod-bottom {
-		@include bottom;
-	}
-
-	&.mod-left {
-		@include left;
-	}
-
-	&.mod-right {
-		@include right;
-	}
-}
