+++ stories/qa/notch-box/notch-box.stories.html
@@ -1,0 +1,34 @@
+<h1>NotchBox</h1>
+
+<!-- Frame structure -->
+<section class="contentSection">
+	<h2>Mods</h2>
+	<div class="u-displayFlex u-gapBigger u-alignItemsFlexStart"
+		style="--component-notchbox-notch-position:50%; --component-notchbox-notch-percent: .5;">
+		<div class="notchBox mod-notchTop">
+			<div class="notchBox-badge"></div>
+			<div class="notchBox-wrapper">
+				<div class="notchBox-wrapper-content">
+					<strong>mod-notchTop</strong>
+					<div>Lorem ipsum dolor sit amet, consectetur adipiscing
+						elit. Maecenas vehicula
+						lacinia nisl in maximus. Maecenas nec enim non nulla faucibus posuere.</div>
+				</div>
+			</div>
+		</div>
+		<div class="notchBox mod-notchLeft">
+			<div class="notchBox-badge"></div>
+			<div class="notchBox-wrapper">
+				<div class="notchBox-wrapper-content">
+					<strong>mod-notchLeft</strong>
+					<div>Lorem ipsum dolor sit amet, consectetur adipiscing
+						elit. Maecenas vehicula
+						lacinia nisl in maximus. Maecenas nec enim non nulla faucibus posuere.</div>
+				</div>
+			</div>
+		</div>
+	</div>
+</section>
+
+<!-- To tell the ui-diff tool that the page has finished rendering -->
+<span id="ready"></span>
