--- stories/qa/notched-box/notched-box.stories.html
@@ -1,48 +1,0 @@
-<h1>NotchedBox</h1>
-
-<!-- Frame structure -->
-<section class="contentSection">
-	<h2>Mods</h2>
-	<div class="u-displayFlex u-gapBig u-alignItemsFlexStart"
-		style="--component-notchedbox-notch-position:50%; --component-notchedbox-notch-percent: .5;">
-		<div class="notchedBox mod-top">
-			<div class="notchedBox-badge"></div>
-			<div class="notchedBox-content">
-				<strong>mod-top</strong>
-				<div>Lorem ipsum dolor sit amet, consectetur adipiscing
-					elit. Maecenas vehicula
-					lacinia nisl in maximus. Maecenas nec enim non nulla faucibus posuere.</div>
-			</div>
-		</div>
-		<div class="notchedBox mod-bottom">
-			<div class="notchedBox-badge"></div>
-			<div class="notchedBox-content">
-				<strong>mod-bottom</strong>
-				<div>Lorem ipsum dolor sit amet, consectetur adipiscing
-					elit. Maecenas vehicula
-					lacinia nisl in maximus. Maecenas nec enim non nulla faucibus posuere.</div>
-			</div>
-		</div>
-		<div class="notchedBox mod-left">
-			<div class="notchedBox-badge"></div>
-			<div class="notchedBox-content">
-				<strong>mod-left</strong>
-				<div>Lorem ipsum dolor sit amet, consectetur adipiscing
-					elit. Maecenas vehicula
-					lacinia nisl in maximus. Maecenas nec enim non nulla faucibus posuere.</div>
-			</div>
-		</div>
-		<div class="notchedBox mod-right">
-			<div class="notchedBox-badge"></div>
-			<div class="notchedBox-content">
-				<strong>mod-right</strong>
-				<div>Lorem ipsum dolor sit amet, consectetur adipiscing
-					elit. Maecenas vehicula
-					lacinia nisl in maximus. Maecenas nec enim non nulla faucibus posuere.</div>
-			</div>
-		</div>
-	</div>
-</section>
-
-<!-- To tell the ui-diff tool that the page has finished rendering -->
-<span id="ready"></span>
